CQB: Randomly distribute unit total amongst areas
Here's an example:
* Unit limit set to 1
* Two game logics with a total of one position each.

With this change in place, the position will be randomly selected
between the two areas. If area A has double amount of positions to that
of area B, B is half as likely to be selected.

If unit limit is set to 10 and there are 10+10 positions available,
there should be around 5 in each area. If there's 5+5 positions, all of
them would be filled like before.

This also means that the numberOfEnemies limit is global. Having
logic-specific limits can come later, if I want them to.