Infidels: Remove
We're removing Massi's mods from our modset and this mission requires
it. I don't want to spend time porting it because it's not even
finished, and with the new 3D editor that is coming soon, making
something like it will be easy when I decide to work on it again.